+++ packages/components/src/components/atoms/form/Expansion.tsx
@@ -10,7 +10,7 @@ import Button from '../buttons/Button'
 import InputText from './InputText'
 import { styled } from '@mui/system'
 import { useTranslation } from 'next-i18next'
-import TextFieldTagsError from './TextFieldTagsError'
+import TextFieldTags from './TextFieldTags'
 
 const CustomPropError = ['error']
 const CustomRoot = styled('div', {
@@ -190,7 +190,7 @@ function Expansion(props: IProps): JSX.Element {
                   }}
                   showError={showError}
                 />
-                <TextFieldTagsError
+                <TextFieldTags
                   required
                   label={t('expansion terms')}
                   fullWidth={fullWidth}
