+++ packages/components/src/components/atoms/form/Range.test.tsx
@@ -2,42 +2,40 @@ import React from 'react'
 
 import { renderWithProviders } from '../../../utils/tests'
 
-import RangeError from './RangeError'
+import Range from './Range'
 import { screen } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 
 describe('RangeError', () => {
   it('sould match snapshot', () => {
     const { container } = renderWithProviders(
-      <RangeError label="Label" value={[1, 3]} />
+      <Range label="Label" value={[1, 3]} />
     )
     expect(container).toMatchSnapshot()
   })
 
   it('should display error with showError prop when the field is required with missing value', () => {
     renderWithProviders(
-      <RangeError label="Label" required showError value={[1, null]} />
+      <Range label="Label" required showError value={[1, null]} />
     )
     expect(screen.getByText('formError.valueMissing')).toBeInTheDocument()
   })
 
   it('should not display error without showError prop when the field is required with missing value', () => {
-    renderWithProviders(
-      <RangeError label="Label" required value={[null, null]} />
-    )
+    renderWithProviders(<Range label="Label" required value={[null, null]} />)
     expect(screen.queryByText('formError.valueMissing')).not.toBeInTheDocument()
   })
 
   it('should not display error with showError prop when the field is required and disabled with missing value', () => {
     renderWithProviders(
-      <RangeError label="Label" required showError disabled value={[null, 1]} />
+      <Range label="Label" required showError disabled value={[null, 1]} />
     )
     expect(screen.queryByText('formError.valueMissing')).not.toBeInTheDocument()
   })
 
   it('shoud display error with showError prop when the field has additional validation rules', () => {
     renderWithProviders(
-      <RangeError
+      <Range
         label="Label"
         showError
         required
@@ -50,7 +48,7 @@ describe('RangeError', () => {
 
   it('should display a replacement error message with replacementErrorsMessages prop when the field has an error', () => {
     renderWithProviders(
-      <RangeError
+      <Range
         label="Label"
         showError
         required
