+++ packages/components/src/components/index.ts
@@ -18,10 +18,12 @@ export { default as InputText } from './atoms/form/InputText'
 export { default as RadioGroupWithoutError } from './atoms/form/RadioGroupWithoutError'
 export { default as RadioGroup } from './atoms/form/RadioGroup'
 export { default as Range } from './atoms/form/Range'
-export { default as RangeError } from './atoms/form/RangeError'
+export { default as RangeWithoutError } from './atoms/form/RangeWithoutError'
 export { default as Slider } from './atoms/form/Slider'
 export { default as Switch } from './atoms/form/Switch'
+export { default as TextareaWithoutError } from './atoms/form/TextareaWithoutError'
 export { default as Textarea } from './atoms/form/Textarea'
+export { default as TreeSelectorWithoutError } from './atoms/form/TreeSelectorWithoutError'
 export { default as TreeSelector } from './atoms/form/TreeSelector'
 export { default as FormatText } from './atoms/format/FormatText'
 export { default as Help } from './atoms/help/Help'
